requirements.txt
PyYAML>=6.0
psutil>=5.9.0
click>=8.0.0
colorama>=0.4.4
